chore: cleanup and folder restructured
